claude-sample-cpy/requirements.txt
functions-framework==3.*
flask
google-cloud-logging
google-cloud-aiplatform>=1.110.0
anthropic[vertex]
markdown
selectolax
